+++ sketch/IMU.h
@@ -1,6 +1,7 @@
 #ifndef IMU_H
 #define IMU_H
 
+#include <Arduino.h>
 #include <Adafruit_BNO08x.h>
 
 class IMU {
@@ -8,12 +9,12 @@ public:
     IMU();
     bool init();
     void update();
-    float get_heading(); // renvoie l'angle fait avec le nord magnétique en degrés
+    uint16_t get_heading(); // renvoie l'angle fait avec le nord magnétique en degrés
 
 private:
     Adafruit_BNO08x bno08x;
     sh2_SensorValue_t sensorValue;
-    float heading;
+    uint16_t heading;
     bool configured;
 };
